add new state `MerelySyncedEnough` for PeerClassification
this is for nodes that can serve up a recent-enough block hash, but
don't meet the more strict uptime-over-time criteria. We try and serve
a couple of these along with nodes that meet all the criteria, in order
to quickly integrate new nodes into the network, while still providing
high-reliability nodes as well.