make probe_schedule_for_node compile but it's still very wrong
the difficulty we're dealing with is how to cleanly handle the
invalidation event that happens when we notice the peer connection
metadata changing

currently we set new_peer_stats.block_probe to its default value
(None for the times, and 0 for the counts), and while this will take
care of taking the node out of the AllGood status, it's unclear
which state this will put the node into *afterwards*.

this is a problem because we do get_classification inside the interval
logic, and if we want to properly handle the block probe invalidation,
we need to know which code path[s] the newly-invalidated node will take.

The three options we see are:

1) set a bespoke flag when we invalidate, and add bespoke logic to
check the flag in the interval logic to ensure the node receives a block
probe immediately.

2) find a way to guarantee the node ends up in MerelySynced or
EventuallyMaybe states, by assuring that MerelySynced's block probe age
threshold and/or that EventuallyMaybe's protocol negotiation age
threshold are lenient enough. This seems ultra-touchy, since those are
likely gonna be user-configurable

3) don't actually change new_peer_stats.block_probe at all! and
alongside that, we set a flag to indicate to both the classification
logic *and* the interval logic that the block probe has been invalidated
and to handle it special-case. This should mean that all the tests will
run as if it was in the AllGood status right up until the block probe
test is run...I think.

I think generally we want to ensure that the different states do not
interact in weird ways with each other, especially since we want the
AllGood logic (the most complex) to be run with opinionated defaults and
MerelySynced and EventuallyMaybe to be more user-configurable.